requirements.txt
fastapi~=0.111.0
pydantic~=2.7.2
passlib~=1.7.4
uvicorn~=0.30.0
tortoise-orm~=0.21.2
asyncpg~=0.29.0
pydantic-settings~=2.2.1
bcrypt~=4.1.3
pyjwt~=2.8.0
aerich~=0.7.2
cryptography~=42.0.7